refactor(*): started splitting files, minor spec fixes, fixed imports
