+++ test/validation/valchecks_test.py
@@ -1,12 +1,10 @@
 import unittest
 from datetime import date
-from typing import List, Callable, Mapping
+from typing import List
 from inspect import isclass
 from parameterized import parameterized
-from otri.filtering.stream import Stream
-from otri.validation.validation import ValidatorFilter
-from otri.validation.valchecks import *
-from otri.validation.exceptions import *
+from otri.validation.valchecks import check_non_null, check_positive, make_check_range, make_check_set
+from otri.validation.exceptions import NullError, RangeError, AtomValueError
 
 
 ex_start_date = date(2020, 4, 10)
